extract re-rendering to message component
