+++ src/LoggerList.tsx
@@ -1,24 +1,13 @@
-import { useEffect, useState } from "react";
-
 import { CreateLogger } from "./CreateLogger";
 import { LoggerListItem } from "./LoggerListItem";
 import { LoggerListNav } from "./LoggerListNav";
 import { Stack } from "@mui/material";
 import { deleteLog } from "./dbManagement";
 import { useLoggerListContext } from "./LoggerListContext";
+import { useState } from "react";
 
 export function LoggerList() {
   const [editMode, setEditMode] = useState(false);
-
-  // re-render the entire app once per minute to update the "last updated mins ago" text
-  const [renderPls, setRenderPls] = useState(0);
-  useEffect(() => {
-    const update = () => {
-      setRenderPls(Date.now());
-    };
-    const timerID = setInterval(update, 60000);
-    return () => clearInterval(timerID);
-  }, [renderPls]);
 
   const toggleEditMode = () => {
     setEditMode((oldMode) => {
